Solve Puzzle9 part 1
Use a LinkedList to record valid sums as we traverse the list of the data only once.
Each time we move forward, pull off the first 25 elements of the list - should be O(n) for n size of preamble (not data!) - and then append the newly generated 25.

Slower to search for a value though, so that's the tradeoff